Refactoring of constants, filters and helpers
