+++ pytrafikverket/__init__.py
@@ -1,6 +1,16 @@
 """Pytrafikverket module."""
 
 # flake8: noqa
+from .filters import (
+    AndFilter,
+    FieldFilter,
+    FieldSort,
+    Filter,
+    FilterOperation,
+    OrFilter,
+    SortOrder,
+)
+from .helpers import NodeHelper
 from .models import (
     CameraInfoModel,
     DeviationInfoModel,
@@ -13,14 +23,6 @@ from .models import (
     WeatherStationInfoModel,
 )
 from .trafikverket import (
-    AndFilter,
-    FieldFilter,
-    FieldSort,
-    Filter,
-    FilterOperation,
-    NodeHelper,
-    OrFilter,
-    SortOrder,
     Trafikverket,
 )
 from .trafikverket_camera import TrafikverketCamera
