+++ pytrafikverket/trafikverket.py
@@ -3,121 +3,20 @@
 from __future__ import annotations
 
 import logging
-from abc import ABCMeta, abstractmethod
-from datetime import datetime
-from enum import Enum
-from typing import Any
 
 import aiohttp
 from lxml import etree
 
+from .const import API_URL
 from .exceptions import InvalidAuthentication, UnknownError
+from .filters import FieldFilter, FieldSort, Filter
+from .helpers import NodeHelper
 
 LOGGER = logging.getLogger(__name__)
 
 
-class FilterOperation(Enum):
-    """Contains all field filter operations."""
-
-    EQUAL = "EQ"
-    EXISTS = "EXISTS"
-    GREATER_THAN = "GT"
-    GREATER_THAN_EQUAL = "GTE"
-    LESS_THAN = "LT"
-    LESS_THAN_EQUAL = "LTE"
-    NOT_EQUAL = "NE"
-    LIKE = "LIKE"
-    NOT_LIKE = "NOTLIKE"
-    #    IN = "IN"  # noqa: ERA001
-    NOT_IN = "NOTIN"  # codespell:ignore
-    WITH_IN = "WITHIN"
-
-
-class SortOrder(Enum):
-    """Specifies how rows of data are sorted."""
-
-    ASCENDING = "asc"
-    DESCENDING = "desc"
-
-
-class FieldSort:
-    """What field and how to sort on it."""
-
-    def __init__(self, field: str, sort_order: SortOrder) -> None:
-        """Initialize the class."""
-        self._field = field
-        self._sort_order = sort_order
-
-    def to_string(self) -> str:
-        """Sort_order as string."""
-        return self._field + " " + self._sort_order.value
-
-
-class Filter:
-    """Base class for all filters."""
-
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def generate_node(self, parent_node: etree._Element) -> Any:
-        """Generate node."""
-
-
-class FieldFilter(Filter):
-    """Used to filter on one field."""
-
-    def __init__(self, operation: FilterOperation, name: str, value: Any) -> None:
-        """Initialize the class."""
-        self.operation = operation
-        self.name = name
-        self.value = value
-
-    def generate_node(self, parent_node: etree._Element) -> Any:
-        """Return element node for field filter."""
-        filter_node: etree._Element = etree.SubElement(
-            parent_node, self.operation.value
-        )
-        filter_node.attrib["name"] = self.name
-        filter_node.attrib["value"] = self.value
-        return filter_node
-
-
-class OrFilter(Filter):
-    """Used to create a Or filter."""
-
-    def __init__(self, filters: list[Filter]) -> None:
-        """Initialize the class."""
-        self.filters = filters
-
-    def generate_node(self, parent_node: Any) -> Any:
-        """Return element node for filter."""
-        or_node: etree._Element = etree.SubElement(parent_node, "OR")
-        for sub_filter in self.filters:
-            sub_filter.generate_node(or_node)
-        return or_node
-
-
-class AndFilter(Filter):
-    """Used to create a And filter."""
-
-    def __init__(self, filters: list[Filter]) -> None:
-        """Initialize the class."""
-        self.filters = filters
-
-    def generate_node(self, parent_node: Any) -> etree._Element:
-        """Return element node for filter."""
-        or_node: etree._Element = etree.SubElement(parent_node, "AND")
-        for sub_filter in self.filters:
-            sub_filter.generate_node(or_node)
-        return or_node
-
-
 class Trafikverket:
     """Class used to communicate with trafikverket api."""
-
-    _api_url = "https://api.trafikinfo.trafikverket.se/v2/data.xml"
-    date_time_format = "%Y-%m-%dT%H:%M:%S.%f%z"
-    date_time_format_for_modified = "%Y-%m-%dT%H:%M:%S.%fZ"
 
     def __init__(self, client_session: aiohttp.ClientSession, api_key: str) -> None:
         """Initialize TrafikInfo object."""
@@ -169,7 +68,7 @@ class Trafikverket:
         LOGGER.debug("Sending query with: %s", request_data_text)
         headers = {"content-type": "text/xml"}
         async with self._client_session.post(
-            Trafikverket._api_url, data=request_data_text, headers=headers
+            API_URL, data=request_data_text, headers=headers
         ) as response:
             content = await response.text()
             LOGGER.debug("Response received with: %s", content)
@@ -201,93 +100,3 @@ class TrafikverketBase:
     def __init__(self, client_session: aiohttp.ClientSession, api_key: str) -> None:
         """Initialize Trafikverket Base."""
         self._api = Trafikverket(client_session, api_key)
-
-
-class NodeHelper:
-    """Helper class to get node content."""
-
-    def __init__(self, node: etree._ElementTree) -> None:
-        """Initialize the class."""
-        self._node = node
-
-    def get_text(self, field: str) -> str | None:
-        """Return the text in 'field' from the node or None if not found."""
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        if len(nodes) == 0:
-            return None
-        if len(nodes) > 1:
-            raise ValueError("Found multiple nodes should only 0 or 1 is allowed")
-        LOGGER.debug("Return text value %s", nodes[0].text)
-        value: str = nodes[0].text
-        return value
-
-    def get_number(self, field: str) -> float | None:
-        """Return the number in 'field' from the node or None if not found."""
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        if len(nodes) == 0:
-            return None
-        if len(nodes) > 1:
-            raise ValueError("Found multiple nodes should only 0 or 1 is allowed")
-        LOGGER.debug("Return number value %s", nodes[0].text)
-        try:
-            value = float(nodes[0].text)
-        except ValueError:
-            return None
-        return value
-
-    def get_texts(self, field: str) -> list[str] | None:
-        """Return a list of texts from the node selected by 'field' or None."""
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        result = []
-        for line in nodes:
-            result.append(line.text)
-        LOGGER.debug("Return texts value %s", result)
-        return result
-
-    def get_datetime_for_modified(self, field: str) -> datetime | None:
-        """Return datetime object from node, selected by 'field'.
-
-        Format of the text is expected to be modifiedTime-format.
-        """
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        if len(nodes) == 0:
-            return None
-        if len(nodes) > 1:
-            raise ValueError("Found multiple nodes should only 0 or 1 is allowed")
-        LOGGER.debug("Return modified value %s", nodes[0].text)
-        return datetime.strptime(
-            nodes[0].text, Trafikverket.date_time_format_for_modified
-        )
-
-    def get_datetime(self, field: str) -> datetime | None:
-        """Return a datetime object from node, selected by 'field'."""
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        if len(nodes) == 0:
-            return None
-        if len(nodes) > 1:
-            raise ValueError("Found multiple nodes should only 0 or 1 is allowed")
-        LOGGER.debug("Return datetime value %s", nodes[0].text)
-        return datetime.strptime(nodes[0].text, Trafikverket.date_time_format)
-
-    def get_bool(self, field: str) -> bool | None:
-        """Return True if value selected by field is 'true' else returns False."""
-        nodes: list[etree._Element] | None = self._node.xpath(field)
-        if nodes is None:
-            return None
-        if len(nodes) == 0:
-            return None
-        if len(nodes) > 1:
-            raise ValueError("Found multiple nodes should only 0 or 1 is allowed")
-        LOGGER.debug("Return bool value %s", nodes[0].text)
-        value: bool = nodes[0].text.lower() == "true"
-        return value
